Implement MAP_DOUBLING fixes 3a/3b + 18 review fixes across back-end/mapping/front-end
Map-doubling plan (docs/MAP_DOUBLING_FIX_PLAN.md, now marked implemented):
- 3a: NSSM global-init yaw bounds clamped to ±min(5·rotation_std,
  nssm/max_yaw_vs_compass) — candidates beyond the compass band could only be
  rejected (56% of CHL_Pool rejects), and the loose ±5σ window is what let the
  Sobol search land in the ~90° alias basin. Translation keeps full ±5σ.
- 3b: new keyframe_min_points (slam.yaml: 10; 0 = off) — near-empty feature
  clouds no longer become 0-point keyframes; DR odometry bridges the gap.

SLAM back-end (post-loop verification + PCM):
- Windowed (W=20) max-RMS replaces the global compass yaw-RMS: a localized
  fold no longer dilutes below a fixed threshold as the mission grows.
- Chain-tear metric is now the body-frame link-vector difference (each chart's
  link rotated into its own keyframe heading): direction-sensitive, so a
  fold-back that REVERSES a link (length preserved, old metric read 0) trips.
- Reverted loop closures are quarantined (ICPResult.rejected): a clique that
  demonstrably bent the graph can neither re-insert nor vote in PCM again,
  ending the accept/revert/rebuild churn; solver-failure rollbacks do not
  quarantine (the loops were never judged).
- PCM chi2 denominator frame fix: GTSAM marginals are body-frame tangent
  covariances — rotate each into world before summing, then into the error
  frame (previously treated as world with a spurious unrotation).
- committed_graph_ is mirrored immediately after a successful solve and
  truncated on revert, and the whole post-solve section is exception-guarded:
  a throw between solve and commit can no longer desync the mirror, strand
  loop marks, or leave half-swept poses (unwinds like a revert, no quarantine).
- All defense thresholds now visible in config/slam.yaml.

Mapping:
- free_tile_min_points judged on the FILTERED cloud (the outlier filter could
  empty a cloud after the raw-count check, stamping the all-free wedge over
  real walls — the exact failure the gate exists to prevent); repo default 5.
- frame_y_sign chirality fix: hit binning uses atan2(frame_y_sign·py, px) and
  the mosaic reads the mirrored column only for +1 — with enu_world=false both
  products were port<->starboard mirrored against the trajectory.
- Tile re-render thresholds tightened (0.5/0.05 -> 0.2/0.02 in mapping.yaml):
  sub-threshold graph corrections were silently absorbed, ghosting up to half
  the ~1 m doubling gap the campaign is chasing.
- stream_grace (1.0 s) separates the conclusively-lost verdict from the 2 ms
  pairing tolerance — a milliseconds-late best-effort ping is no longer
  irreversibly skipped; dead/misrouted streams now warn instead of silently
  publishing an empty map.
- Per-keyframe pairing tolerance scales with the float32 stamp-offset
  quantization (fixed 2 ms starts spuriously skipping keyframes ~4 h in).
- Submap tiles store intensity as uint8 (sums stay uint32 in the global
  grids): 4x less per-keyframe memory on the dominant allocation.
- kf.k records the actual slot when padding fires; mirror-column queries are
  clamped into the aperture (asymmetric fans no longer alias to column 0).

Front-end / build:
- tilt_at() clock-domain guard: tilt samples are arrival-stamped (unstamped
  Float32) while pings carry driver stamps — when the domains disagree (bag
  replay without sim time) the edge clamp pinned every lookup to the OLDEST
  buffered sample; now falls back to the newest (the pre-existing latch
  behavior) with a throttled warning. Also covers a stalled tilt stream.
- CUDA arch self-heal restricted to the actual stale legacy default (<= sm_52):
  a deliberate -DCMAKE_CUDA_ARCHITECTURES=70/72 (V100/Xavier) is kept with a
  warning instead of silently replaced with sm_86 cubins that cannot load.
- feature.yaml documents apply_head_tilt / max_head_pitch / map-stream keys.

Verified: full colcon build with CUDA ON (the arch change live-fired on this
workspace's real stale '52' cache and reset it), all changed sources compile
standard-only, and every touched yaml parses with the expected keys.

+++ include/sonar_slam_cpp/mapping.hpp
@@ -51,16 +51,19 @@ struct Submap
   // y-left), shared across keyframes that share the sonar geometry (rebuilt
   // only when the range/aperture changes — mapping.py reuses the previous).
   std::shared_ptr<const Matrix> sonar_xy;
-  // Per-pixel deposits (aligned with sonar_xy rows).
+  // Per-pixel deposits (aligned with sonar_xy rows). Intensity is a raw
+  // 0..255 pixel per cell — uint8 (the uint32 SUMS live only in the global
+  // grids); at ~1 tile/keyframe this is a 4x memory saving on the dominant
+  // per-keyframe allocation of a long mission.
   std::vector<float> logodds;      // occupancy logodds
-  std::vector<uint32_t> intensity;  // raw backscatter 0..255
+  std::vector<uint8_t> intensity;  // raw backscatter 0..255
 
   // Current placement into the global grid (recomputed by fit_grid). r/c are
   // cell indices; l/i are the deduplicated per-cell deposits actually summed
   // into the grids by inc_grid (and subtracted by dec_grid).
   std::vector<int> r, c;
   std::vector<float> l;
-  std::vector<uint32_t> i;
+  std::vector<uint8_t> i;
 };
 
 class Mapping
